refactor: 비즈니스 로직을 서비스 모듈로 분리
- 할인, 포인트, 데이터 변환 로직을 별도 서비스로 모듈화
- main.basic.js 코드 정리 및 테스트 에러 수정
- productStore 의존성 주입 문제 해결
