+++ src/basic/features/data/index.js
@@ -1,0 +1,2 @@
+// 🏪 데이터 서비스 exports
+export { createInitialProductState, transformServerDataToClientState } from './dataService.js';
