+++ src/basic/features/discount/index.js
@@ -1,0 +1,2 @@
+// 🏪 할인 서비스 exports
+export { discountService } from './discountService.js';
