+++ src/basic/features/index.js
@@ -1,6 +1,9 @@
 // 🏪 기능 모듈 exports
 export * from './cart/index.js';
+export * from './data/index.js';
+export * from './discount/index.js';
 export * from './events/index.js';
 export { setupObservers } from './observerFactory.js';
+export * from './points/index.js';
 export * from './product/index.js';
 export * from './ui/index.js';
