+++ src/basic/features/points/index.js
@@ -1,0 +1,2 @@
+// 🏪 포인트 서비스 exports
+export { pointService } from './pointService.js';
